Rename it to `simple_optimization`
