+++ compiler/noirc_evaluator/src/ssa/opt/brillig_array_gets.rs
@@ -37,7 +37,7 @@ impl Ssa {
 
 impl Function {
     pub(super) fn brillig_array_gets(&mut self) {
-        self.mutate(|context| {
+        self.simple_optimization(|context| {
             let instruction_id = context.instruction_id;
             let instruction = context.instruction();
             if let Instruction::ArrayGet { array, index } = instruction {
