+++ compiler/noirc_evaluator/src/ssa/opt/mod.rs
@@ -29,6 +29,7 @@ mod remove_enable_side_effects;
 mod remove_if_else;
 mod remove_truncate_after_range_check;
 mod remove_unreachable;
+mod simple_optimization;
 mod simplify_cfg;
 mod unrolling;
